Add files via upload
This code is for Julie Luscomb's C programming class, CSCI 2473

Book: C Primer Plus (Developer's Library) 6th Edition
by Stephen Prata 

Using the Driver Program for Homework; create, test, and turn in the ASCII code to character function - #2 on page 97. 
Using the Driver Program for Homework; create, test, and turn in the conversion function - #8 on page 97. This should be added to the single file used for the first part of this assignment. Include your answer to the question as a comment at the top of this function.